[iOS] Introduce static library for generic parts
Works as intended, but needs to be completed. 1 further todo remains,
see below.

Adapt generation of project file as necessary to include library.
Generated import statements need to include path to library.

Initial version of library contains one class: DataTransferObject (which
hence has been removed from resources and FileStructure). Adapt import
statements of static, but not yet migrated files (to be undone when they
are also moved to library).

TODO: the actual Md2Library folder still needs to be copied to the
result folder manually. This needs to be automated as well.